MediaRouter: make request id globally unique
A unique request id can be used to identify each request
and the client of the request.

Test: atest mediaroutertest

+++ media/tests/MediaRouteProvider/src/com/android/mediarouteprovider/example/SampleMediaRoute2ProviderService.java
@@ -153,7 +153,7 @@ public class SampleMediaRoute2ProviderService extends MediaRoute2ProviderService
 
     @Override
     public void onCreateSession(String packageName, String routeId, String controlCategory,
-            int requestId) {
+            long requestId) {
         MediaRoute2Info route = mRoutes.get(routeId);
         if (route == null || TextUtils.equals(ROUTE_ID3_SESSION_CREATION_FAILED, routeId)) {
             // Tell the router that session cannot be created by passing null as sessionInfo.
